refactoring

+++ src/main/java/repository/MatchRepository.java
@@ -1,13 +1,13 @@
 package repository;
 
-import entity.Match;
+import model.Match;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.Session;
 import java.util.List;
 import java.util.Optional;
 
 @RequiredArgsConstructor
-public class MatchRepository implements Repository<Match, Long> {
+public class MatchRepository implements Repository<Match, Integer> {
 
     @Override
     public Optional<Match> save(Match match, Session session) {
@@ -16,7 +16,7 @@ public class MatchRepository implements Repository<Match, Long> {
     }
 
     @Override
-    public Optional<Match> getById(Long id, Session session) {
+    public Optional<Match> getById(Integer id, Session session) {
         return Optional.ofNullable(session.get(Match.class, id));
     }
